reorganize mail package
