+++ app/src/main/java/edu/kmaooad/capstone23/mail/service/Notification.java
@@ -1,4 +1,4 @@
-package edu.kmaooad.capstone23.mail;
+package edu.kmaooad.capstone23.mail.service;
 
 public class Notification {
     private final String email;
