+++ app/src/main/java/edu/kmaooad/capstone23/students/handlers/NotifyStudentHandler.java
@@ -3,11 +3,11 @@ package edu.kmaooad.capstone23.students.handlers;
 import edu.kmaooad.capstone23.common.CommandHandler;
 import edu.kmaooad.capstone23.common.ErrorCode;
 import edu.kmaooad.capstone23.common.Result;
-import edu.kmaooad.capstone23.mail.NotificationMailService;
+import edu.kmaooad.capstone23.mail.service.NotificationMailService;
 import edu.kmaooad.capstone23.students.commands.notifications.NotifyStudent;
 import edu.kmaooad.capstone23.students.events.StudentNotified;
-import edu.kmaooad.capstone23.mail.Notification;
-import edu.kmaooad.capstone23.mail.NotificationBuilder;
+import edu.kmaooad.capstone23.mail.service.Notification;
+import edu.kmaooad.capstone23.mail.service.NotificationBuilder;
 import jakarta.enterprise.context.RequestScoped;
 import jakarta.inject.Inject;
 import org.bson.types.ObjectId;
